requirements.txt
langchain-ollama
langchain
python-dotenv
Pypdf2
streamlit
streamlit_chat
faiss-cpu